@@ -34,29 +34,32 @@
     "@angular/common": "^16.1.1",
     "@angular/compiler": "^16.1.1",
     "@angular/core": "^16.1.1",
-    "@angular/forms": "^16.1.5",
-    "@angular/localize": "^16.1.5",
+    "@angular/forms": "^16.2.2",
+    "@angular/localize": "^16.2.2",
     "@angular/material": "^16.1.1",
-    "@angular/material-moment-adapter": "^16.1.5",
+    "@angular/material-moment-adapter": "^16.2.1",
     "@angular/platform-browser": "^16.1.1",
-    "@angular/platform-browser-dynamic": "^16.1.5",
-    "@angular/router": "^16.1.5",
+    "@angular/platform-browser-dynamic": "^16.2.2",
+    "@angular/router": "^16.2.2",
     "@dlr-shepard/shepard-client": "2023.4.11",
-    "@supabase/supabase-js": "^2.26.0",
-    "@swimlane/ngx-graph": "^8.2.1",
+    "@supabase/supabase-js": "^2.33.1",
+    "@swimlane/ngx-graph": "^8.2.2",
+    "buffer": "^6.0.3",
     "cross-env": "7.0.3",
+    "csv-parse": "^5.5.0",
+    "jszip": "^3.10.1",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
     "node-sass-json-importer": "^4.3.0",
     "rxjs": "~7.8.0",
-    "sass": "^1.63.6",
+    "sass": "^1.66.1",
     "snakecase-keys": "^5.4.6",
-    "tslib": "^2.6.0",
-    "yaml": "^2.3.1",
+    "tslib": "^2.6.2",
+    "yaml": "^2.3.2",
     "zone.js": "^0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.4",
+    "@angular-devkit/build-angular": "^16.2.0",
     "@angular-eslint/builder": "16.0.3",
     "@angular-eslint/eslint-plugin": "16.0.3",
     "@angular-eslint/eslint-plugin-template": "16.0.3",
@@ -66,13 +69,14 @@
     "@angular/compiler-cli": "^16.1.1",
     "@openapitools/openapi-generator-cli": "^2.7.0",
     "@types/jasmine": "~4.3.1",
-    "@types/lodash": "^4.14.195",
+    "@types/jszip": "^3.4.1",
+    "@types/lodash": "^4.14.197",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "cypress": "^12.17.1",
-    "eslint": "^8.45.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
+    "cypress": "^12.17.4",
+    "eslint": "^8.48.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-jsdoc": "^40.3.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "eslint-plugin-rxjs-angular": "^2.0.1",
