@@ -10,6 +10,7 @@ import {
   mergeMap,
   Observable,
   Subject,
+  switchMap,
   take,
   takeUntil,
 } from 'rxjs';
@@ -21,6 +22,7 @@ import { ProcessFacadeService } from 'src/app/services/process-facade.service';
 import { StepFacadeService } from 'src/app/services/step-facade.service';
 import { elementComparator } from '../../shared/util/common-utils';
 import { ElementViewModelFacadeService } from '../../services';
+import { SerializationService } from '../../services/serialization.service';
 
 @Component({
   selector: 'app-process-detail',
@@ -30,7 +32,7 @@ import { ElementViewModelFacadeService } from '../../services';
 })
 export class ProcessDetailComponent implements OnDestroy {
   public title = '';
-  public currentStepId: number | null | undefined;
+  public currentStepId: number | null = 0;
   public breadcrumbs: Breadcrumb[] = [];
   public steps: Step[] = [];
   public stepTitles: string[] = [];
@@ -49,10 +51,12 @@ export class ProcessDetailComponent implements OnDestroy {
   public tableData$: Observable<any> = new Observable<any>();
 
   public currentIndex = 0;
+  protected loading = false;
 
   private readonly _destroy$: Subject<void> = new Subject<void>();
 
   constructor(
+    private readonly serializationService: SerializationService,
     private readonly processService: ProcessFacadeService,
     private readonly elementService: ElementFacadeService,
     private readonly stepService: StepFacadeService,
@@ -96,7 +100,9 @@ export class ProcessDetailComponent implements OnDestroy {
   public changeContent(index: number): void {
     this.currentIndex = index;
     this.currentStepId =
-      index === this.steps.length ? null : this.steps[this.currentIndex]?.id;
+      index === this.steps.length
+        ? null
+        : this.steps[this.currentIndex].id ?? 0;
     this.processId$()
       .pipe(
         concatMap(
@@ -175,12 +181,31 @@ export class ProcessDetailComponent implements OnDestroy {
     );
   }
 
+  protected downloadDataset(): void {
+    this.loading = true;
+    this.processId$()
+      .pipe(
+        switchMap(id =>
+          this.serializationService.export$(id, this.currentStepId)
+        ),
+        take(1)
+      )
+      .subscribe(blob => {
+        const atag = document.createElement('a');
+        atag.href = URL.createObjectURL(blob);
+        atag.download = 'dataset.zip';
+        document.body.appendChild(atag);
+        atag.click();
+        document.body.removeChild(atag);
+        this.loading = false;
+      });
+  }
+
   private processId$(): Observable<number> {
     return this.activatedRoute.paramMap.pipe(
       distinctUntilChanged(),
       filter(param => !!param.get('processId')),
-      map(param => +param.get('processId')!),
-      takeUntil(this._destroy$)
+      map(param => +param.get('processId')!)
     );
   }
 }
